+++ examples/seismic/tti/tti_example3D.py
